pyproject.toml
[tool.poetry]
name = "llm-search-judge"
version = "0.1.0"
description = ""
authors = ["Doug Turnbull <softwaredoug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12.3,<3.13"
pandas = "^2.2.3"
mlx = "^0.22.0"
mlx-lm = "^0.21.0"
prompt-toolkit = "^3.0.48"
scikit-learn = "^1.6.1"
pyarrow = "^19.0.0"
mlx-vlm = "^0.1.11"
pillow = "^11.1.0"
graphviz = "^0.20.3"
google-cloud-secret-manager = "^2.22.1"
liaison-client = {git = "ssh://git@github.com/dahlialabs/liaison.git", subdirectory = "sdks/python"}
jupyter = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
